# Use std::arch SIMD and runtime target feature detection