--- a/LICENSE-MIT
+++ b/LICENSE-MIT
@@ -1,4 +1,4 @@
-Copyright (c) 2015
+Copyright (c) 2016 - 2018 Ulrik Sverdrup "bluss"
 
 Permission is hereby granted, free of charge, to any
 person obtaining a copy of this software and associated