--- a/src/aligned_alloc.rs
+++ b/src/aligned_alloc.rs
@@ -1,4 +1,5 @@
 
+use std::alloc;
 use std::alloc::{Layout, handle_alloc_error};
 use std::{mem, cmp};
 
@@ -18,7 +19,7 @@ impl<T> Alloc<T> {
         #[cfg(not(debug_assertions))]
         let layout = Layout::from_size_align_unchecked(mem::size_of::<T>() * nelem, align);
         dprint!("Allocating nelem={}, layout={:?}", nelem, layout);
-        let ptr = std::alloc::alloc(layout);
+        let ptr = alloc::alloc(layout);
         if ptr.is_null() {
             handle_alloc_error(layout);
         }
@@ -46,7 +47,7 @@ impl<T> Drop for Alloc<T> {
     fn drop(&mut self) {
         unsafe {
             let layout = Layout::from_size_align_unchecked(mem::size_of::<T>() * self.len, self.align);
-            std::alloc::dealloc(self.ptr as _, layout);
+            alloc::dealloc(self.ptr as _, layout);
         }
     }
 }