--- a/src/dgemm_kernel.rs
+++ b/src/dgemm_kernel.rs
@@ -72,6 +72,40 @@ impl GemmKernel for Gemm {
 pub unsafe fn kernel(k: usize, alpha: T, a: *const T, b: *const T,
                      beta: T, c: *mut T, rsc: isize, csc: isize)
 {
+    // dispatch to specific compiled versions
+    #[cfg(any(target_arch="x86", target_arch="x86_64"))]
+    {
+        if is_x86_feature_detected!("avx") {
+            return kernel_target_avx(k, alpha, a, b, beta, c, rsc, csc);
+        } else if is_x86_feature_detected!("sse") {
+            return kernel_target_sse(k, alpha, a, b, beta, c, rsc, csc);
+        }
+    }
+    kernel_fallback_impl(k, alpha, a, b, beta, c, rsc, csc);
+}
+
+#[inline]
+#[target_feature(enable="avx")]
+#[cfg(any(target_arch="x86", target_arch="x86_64"))]
+pub unsafe fn kernel_target_avx(k: usize, alpha: T, a: *const T, b: *const T,
+                         beta: T, c: *mut T, rsc: isize, csc: isize)
+{
+    kernel_fallback_impl(k, alpha, a, b, beta, c, rsc, csc)
+}
+
+#[inline]
+#[target_feature(enable="sse")]
+#[cfg(any(target_arch="x86", target_arch="x86_64"))]
+pub unsafe fn kernel_target_sse(k: usize, alpha: T, a: *const T, b: *const T,
+                          beta: T, c: *mut T, rsc: isize, csc: isize)
+{
+    kernel_fallback_impl(k, alpha, a, b, beta, c, rsc, csc)
+}
+
+#[inline(always)]
+pub unsafe fn kernel_fallback_impl(k: usize, alpha: T, a: *const T, b: *const T,
+                                   beta: T, c: *mut T, rsc: isize, csc: isize)
+{
     // using `uninitialized` is a workaround for issue https://github.com/bluss/matrixmultiply/issues/9
     let mut ab: [[T; NR]; MR] = ::std::mem::uninitialized();
     let mut a = a;