Improve performance out of the box on x86-64 by using `target_feature`. :1234: :fire: 

Implement a fast 8x8 sgemm kernel for x86-64 using `std::arch` intrinsics. We maintain portability with other platforms.

1. Multiple target_feature invocations of the same implementation allow us to compile multiple versions of the code and jump into the best one at runtime. The same simd intrinsics or Rust code would for example compile to better code inside a `target_feature(enabled="avx")` function than an sse function.

2. Port a BLIS sgemm 8x8 x86 avx kernel to Rust instrinsics. Use the same memory layout in the main loop.

3. Use std::alloc for aligned allocation

4. Runtime target feature detection and target feature specific compilation allows us to offer native performance out of the box on x86-64 (no compiler flags needed)

5. Test compile to multiple targets (x86-64, x86, aarch64) and run the x86 ones in travis

6. in the f32 avx kernel, schedule loads one iteration ahead, this improves throughput

Fixes #14