@bluss 2018-11-12
Benchmarks. Comparing target-feature=+avx compiled from current master with "generic" build on this PR, and we still have improvements! This is tested on my AVX enabled sandy bridge laptop.

We see some overhead in the small sizes for sgemm - because sgemm is now requesting a 32 byte-aligned packing buffer from the system allocator, apparenly this goes through posix_memalign and is a bit slow.

```
before: master native-cpu build
after: this pr generic build

 name               63 ns/iter  62 ns/iter  diff ns/iter   diff % 
 mat_mul_f32::m004  138         231                   93   67.39% 
 mat_mul_f32::m007  169         271                  102   60.36% 
 mat_mul_f32::m008  180         218                   38   21.11% 
 mat_mul_f32::m009  371         489                  118   31.81% 
 mat_mul_f32::m012  458         586                  128   27.95% 
 mat_mul_f32::m016  601         544                  -57   -9.48% 
 mat_mul_f32::m032  2,984       2,451               -533  -17.86% 
 mat_mul_f32::m064  18,254      15,384            -2,870  -15.72% 
 mat_mul_f32::m127  123,472     109,885          -13,587  -11.00% 
 mat_mul_f64::m004  134         130                   -4   -2.99% 
 mat_mul_f64::m007  208         210                    2    0.96% 
 mat_mul_f64::m008  220         230                   10    4.55% 
 mat_mul_f64::m009  417         424                    7    1.68% 
 mat_mul_f64::m012  513         550                   37    7.21% 
 mat_mul_f64::m016  789         827                   38    4.82% 
 mat_mul_f64::m032  4,395       4,347                -48   -1.09% 
 mat_mul_f64::m064  28,604      28,099              -505   -1.77% 
 mat_mul_f64::m127  207,470     199,336           -8,134   -3.92% 
```

The most exciting part is having these gains in a "generic" build where target feature detection takes care of it. The improvement compared with master's generic build is quite big:

```
before: master generic build
after: this pr generic build
 mat_mul_f32::m127  200,461     109,885          -90,576  -45.18% 
 mat_mul_f64::m127  370,069     199,336         -170,733  -46.14%
```