@bluss 2018-11-13
Improved benchmarks for sgemm.

Fixed inflated allocation size bug and brought down the overhead on small matrixes.
Scheduling data loads early got another 10% speedup in the bigger matrices.

Incremental improvents from the previous benchmark to this state of the pr:

```
before: this pr generic build  (previously posted benchmark)
after: this pr generic build

 name               63 ns/iter  62 ns/iter  diff ns/iter   diff % 
 mat_mul_f32::m004  231         191                  -40  -17.32% 
 mat_mul_f32::m008  218         200                  -18   -8.26% 
 mat_mul_f32::m012  586         602                   16    2.73% 
 mat_mul_f32::m016  544         526                  -18   -3.31% 
 mat_mul_f32::m032  2,451       2,266               -185   -7.55% 
 mat_mul_f32::m064  15,384      13,635            -1,749  -11.37% 
 mat_mul_f32::m127  109,885     97,045           -12,840  -11.68%
```

Benchmarks compared with master:

```
before: master native-cpu build
after: this pr generic build  (pr at this time)

 name               63 ns/iter  62 ns/iter  diff ns/iter   diff % 
 mat_mul_f32::m004  138         191                   53   38.41% 
 mat_mul_f32::m008  180         200                   20   11.11% 
 mat_mul_f32::m012  458         602                  144   31.44% 
 mat_mul_f32::m016  601         526                  -75  -12.48% 
 mat_mul_f32::m032  2,984       2,266               -718  -24.06% 
 mat_mul_f32::m064  18,254      13,635            -4,619  -25.30% 
 mat_mul_f32::m127  123,472     97,045           -26,427  -21.40%
```