@bluss 2018-11-15
Benchmarks vs openBLAS's single threaded mode, and the new sgemm (f32) kernel wins. Which is pretty shocking! For the f64 kernel we "compete".

```
test blas_mat_mul_f32::m064 ... bench:      17,496 ns/iter (+/- 322)
test blas_mat_mul_f32::m127 ... bench:     126,223 ns/iter (+/- 4,609)
test blas_mat_mul_f64::m064 ... bench:      25,466 ns/iter (+/- 319)
test blas_mat_mul_f64::m127 ... bench:     188,728 ns/iter (+/- 24,764)
test mat_mul_f32::m064      ... bench:      14,021 ns/iter (+/- 49)
test mat_mul_f32::m127      ... bench:      98,657 ns/iter (+/- 2,477)
test mat_mul_f64::m064      ... bench:      27,382 ns/iter (+/- 2,574)
test mat_mul_f64::m127      ... bench:     204,120 ns/iter (+/- 4,243)
```